refactor(nms): Consolidate CbsdContext
