+++ nms/app/components/lte/LteContext.tsx
@@ -11,7 +11,6 @@
  * limitations under the License.
  */
 import * as React from 'react';
-import CbsdContext from '../context/CbsdContext';
 import InitSubscriberState, {
   FetchSubscriberState,
   getGatewaySubscriberMap,
@@ -20,165 +19,33 @@ import InitSubscriberState, {
 import LoadingFiller from '../LoadingFiller';
 import SubscriberContext from '../context/SubscriberContext';
 import {ApnContextProvider} from '../context/ApnContext';
+import {CbsdContextProvider} from '../context/CbsdContext';
 import {EnodebContextProvider} from '../context/EnodebContext';
-import {GatewayContextProvider} from '../context/GatewayContext';
-import {GatewayPoolsContextProvider} from '../context/GatewayPoolsContext';
-import {GatewayTierContextProvider} from '../context/GatewayTierContext';
-import {LteNetworkContextProvider} from '../context/LteNetworkContext';
-import {PolicyProvider} from '../context/PolicyContext';
-import {TraceContextProvider} from '../context/TraceContext';
-import {useCallback, useEffect, useMemo, useState} from 'react';
-import type {
-  MutableCbsd,
-  MutableSubscriber,
-  PaginatedCbsds,
-  Subscriber,
-  SubscriberState,
-} from '../../../generated';
-
-import * as cbsdState from '../../state/lte/CbsdState';
 import {
   FEG_LTE,
   LTE,
   NetworkId,
   SubscriberId,
 } from '../../../shared/types/network';
+import {GatewayContextProvider} from '../context/GatewayContext';
+import {GatewayPoolsContextProvider} from '../context/GatewayPoolsContext';
+import {GatewayTierContextProvider} from '../context/GatewayTierContext';
+import {LteNetworkContextProvider} from '../context/LteNetworkContext';
+import {PolicyProvider} from '../context/PolicyContext';
+import {TraceContextProvider} from '../context/TraceContext';
+import {useEffect, useState} from 'react';
 import {useEnqueueSnackbar} from '../../hooks/useSnackbar';
+import type {
+  MutableSubscriber,
+  Subscriber,
+  SubscriberState,
+} from '../../../generated';
 
 type Props = {
   networkId: NetworkId;
   networkType: string;
   children: React.ReactNode;
 };
-
-export function CbsdContextProvider({networkId, children}: Props) {
-  const enqueueSnackbar = useEnqueueSnackbar();
-
-  const [isLoading, setIsLoading] = useState(false);
-  const [fetchResponse, setFetchResponse] = useState<PaginatedCbsds>({
-    cbsds: [],
-    total_count: 0,
-  });
-  const [paginationOptions, setPaginationOptions] = useState<{
-    page: number;
-    pageSize: number;
-  }>({
-    page: 0,
-    pageSize: 10,
-  });
-
-  const refetch = useCallback(() => {
-    return cbsdState.fetch({
-      networkId,
-      page: paginationOptions.page,
-      pageSize: paginationOptions.pageSize,
-      setIsLoading,
-      setFetchResponse,
-      enqueueSnackbar,
-    });
-  }, [
-    networkId,
-    paginationOptions.page,
-    paginationOptions.pageSize,
-    setIsLoading,
-    setFetchResponse,
-    enqueueSnackbar,
-  ]);
-
-  useEffect(() => {
-    void refetch();
-  }, [refetch, paginationOptions.page, paginationOptions.pageSize]);
-
-  const state = useMemo(() => {
-    return {
-      isLoading,
-      cbsds: fetchResponse.cbsds,
-      totalCount: fetchResponse.total_count,
-      page: paginationOptions.page,
-      pageSize: paginationOptions.pageSize,
-    };
-  }, [
-    isLoading,
-    fetchResponse.cbsds,
-    fetchResponse.total_count,
-    paginationOptions.page,
-    paginationOptions.pageSize,
-  ]);
-
-  return (
-    <CbsdContext.Provider
-      value={{
-        state,
-        setPaginationOptions,
-        refetch,
-        create: (newCbsd: MutableCbsd) => {
-          return cbsdState
-            .create({
-              networkId,
-              newCbsd,
-            })
-            .catch(e => {
-              enqueueSnackbar?.('failed to create CBSD', {
-                variant: 'error',
-              });
-              throw e as Error;
-            })
-            .then(() => {
-              void refetch();
-            });
-        },
-        update: (id: number, cbsd: MutableCbsd) => {
-          return cbsdState
-            .update({
-              networkId,
-              id,
-              cbsd,
-            })
-            .catch(e => {
-              enqueueSnackbar?.('failed to update CBSD', {
-                variant: 'error',
-              });
-              throw e as Error;
-            })
-            .then(() => {
-              void refetch();
-            });
-        },
-        deregister: (id: number) => {
-          return cbsdState
-            .deregister({
-              networkId,
-              id,
-            })
-            .catch(() => {
-              enqueueSnackbar?.('failed to deregister CBSD', {
-                variant: 'error',
-              });
-            })
-            .then(() => {
-              void refetch();
-            });
-        },
-        remove: (id: number) => {
-          return cbsdState
-            .remove({
-              networkId,
-              cbsdId: id,
-            })
-            .catch(() => {
-              enqueueSnackbar?.('failed to remove CBSD', {
-                variant: 'error',
-              });
-            })
-            .then(() => {
-              void refetch();
-            });
-        },
-      }}>
-      {children}
-    </CbsdContext.Provider>
-  );
-}
 
 export function SubscriberContextProvider(props: Props) {
   const {networkId} = props;
